renamed consts folder to constants
