+++ cspell.json
@@ -5,7 +5,6 @@
   "dictionaries": [],
   "words": [
     "appbar",
-    "consts",
     "QCQKTDFT",
     "randomisers",
     "Swipeable",
