+++ src/components/StepMenu.tsx
@@ -10,8 +10,8 @@ import Typography from '@mui/material/Typography';
 import { useContext, useEffect, useState } from 'react';
 import { FillOnBar } from '../App';
 import '../App.css';
-import { BeatsPerBar } from '../consts/beatsPerBar';
-import { MeasureDivision } from '../consts/measureDivisions';
+import { BeatsPerBar } from '../constsants/beatsPerBar';
+import { MeasureDivision } from '../constsants/measureDivisions';
 import useLocalStorage from '../hooks/useLocalStorage';
 import assertUnreachable from '../utils/assertUnreachable';
 import PatternMaker, { DEFAULT_TIME_SIGNATURE } from '../utils/classes/patternMaker';
