+++ src/components/StepSetup.tsx
@@ -2,10 +2,10 @@ import RefreshIcon from '@mui/icons-material/RefreshRounded';
 import { Button, Typography } from '@mui/material';
 import { useContext, useState } from 'react';
 import { FillOnBar } from '../App';
-import beatIdeas from '../consts/forRandomiser/beatIdeas';
-import fillStarts from '../consts/forRandomiser/fillStarts';
-import fills from '../consts/forRandomiser/fillTypes';
-import { RandomiseMe } from '../consts/forRandomiser/randomiseMe';
+import beatIdeas from '../constsants/forRandomiser/beatIdeas';
+import fillStarts from '../constsants/forRandomiser/fillStarts';
+import fills from '../constsants/forRandomiser/fillTypes';
+import { RandomiseMe } from '../constsants/forRandomiser/randomiseMe';
 import PatternMaker from '../utils/classes/patternMaker';
 import getStringArrayBetweenTwoValues from '../utils/getArrayBetweenValues';
 import Randomiser from './Randomiser';
